+++ pyrealm/demography/crown.py
@@ -14,9 +14,9 @@ from pyrealm.demography.t_model_functions import StemAllometry
 
 
 def _validate_z_qz_args(
-    z: NDArray[np.float32],
-    stem_properties: list[NDArray[np.float32]],
-    q_z: NDArray[np.float32] | None = None,
+    z: NDArray[np.float64],
+    stem_properties: list[NDArray[np.float64]],
+    q_z: NDArray[np.float64] | None = None,
 ) -> None:
     """Shared validation of for crown function arguments.
 
@@ -95,12 +95,12 @@ def _validate_z_qz_args(
 
 
 def calculate_relative_crown_radius_at_z(
-    z: NDArray[np.float32],
-    stem_height: NDArray[np.float32],
-    m: NDArray[np.float32],
-    n: NDArray[np.float32],
+    z: NDArray[np.float64],
+    stem_height: NDArray[np.float64],
+    m: NDArray[np.float64],
+    n: NDArray[np.float64],
     validate: bool = True,
-) -> NDArray[np.float32]:
+) -> NDArray[np.float64]:
     r"""Calculate relative crown radius at a given height.
 
     The crown shape parameters ``m`` and ``n`` define the vertical distribution of
@@ -143,10 +143,10 @@ def calculate_relative_crown_radius_at_z(
 
 
 def calculate_crown_radius(
-    q_z: NDArray[np.float32],
-    r0: NDArray[np.float32],
+    q_z: NDArray[np.float64],
+    r0: NDArray[np.float64],
     validate: bool = True,
-) -> NDArray[np.float32]:
+) -> NDArray[np.float64]:
     r"""Calculate crown radius from relative crown radius and crown r0.
 
     The relative crown radius (:math:`q(z)`) at a given height :math:`z` describes the
@@ -175,14 +175,14 @@ def calculate_crown_radius(
 
 
 def calculate_stem_projected_crown_area_at_z(
-    z: NDArray[np.float32],
-    q_z: NDArray[np.float32],
-    stem_height: NDArray[np.float32],
-    crown_area: NDArray[np.float32],
-    q_m: NDArray[np.float32],
-    z_max: NDArray[np.float32],
+    z: NDArray[np.float64],
+    q_z: NDArray[np.float64],
+    stem_height: NDArray[np.float64],
+    crown_area: NDArray[np.float64],
+    q_m: NDArray[np.float64],
+    z_max: NDArray[np.float64],
     validate: bool = True,
-) -> NDArray[np.float32]:
+) -> NDArray[np.float64]:
     """Calculate stem projected crown area above a given height.
 
     This function calculates the projected crown area of a set of stems with given
@@ -221,15 +221,15 @@ def calculate_stem_projected_crown_area_at_z(
 
 
 def calculate_stem_projected_leaf_area_at_z(
-    z: NDArray[np.float32],
-    q_z: NDArray[np.float32],
-    stem_height: NDArray[np.float32],
-    crown_area: NDArray[np.float32],
-    f_g: NDArray[np.float32],
-    q_m: NDArray[np.float32],
-    z_max: NDArray[np.float32],
+    z: NDArray[np.float64],
+    q_z: NDArray[np.float64],
+    stem_height: NDArray[np.float64],
+    crown_area: NDArray[np.float64],
+    f_g: NDArray[np.float64],
+    q_m: NDArray[np.float64],
+    z_max: NDArray[np.float64],
     validate: bool = True,
-) -> NDArray[np.float32]:
+) -> NDArray[np.float64]:
     """Calculate projected leaf area above a given height.
 
     This function calculates the projected leaf area of a set of stems with given
@@ -328,16 +328,16 @@ class CrownProfile:
     """A Flora or StemTraits instance providing plant functional trait data."""
     stem_allometry: InitVar[StemAllometry]
     """A StemAllometry instance setting the stem allometries for the crown profile."""
-    z: NDArray[np.float32]
+    z: NDArray[np.float64]
     """An array of vertical height values at which to calculate crown profiles."""
 
-    relative_crown_radius: NDArray[np.float32] = field(init=False)
+    relative_crown_radius: NDArray[np.float64] = field(init=False)
     """An array of the relative crown radius of stems at z heights"""
-    crown_radius: NDArray[np.float32] = field(init=False)
+    crown_radius: NDArray[np.float64] = field(init=False)
     """An array of the actual crown radius of stems at z heights"""
-    projected_crown_area: NDArray[np.float32] = field(init=False)
+    projected_crown_area: NDArray[np.float64] = field(init=False)
     """An array of the projected crown area of stems at z heights"""
-    projected_leaf_area: NDArray[np.float32] = field(init=False)
+    projected_leaf_area: NDArray[np.float64] = field(init=False)
     """An array of the projected leaf area of stems at z heights"""
 
     # Information attributes
